Adding on the Bombs, Shooting, and Checkpoint Functionality
Player Character can summon bombs by pressing Y, these bombs only destroy actors with the Enemy Tag and applies knockback to the player character

Added shooting functionality, player can press G and a projectile will be summoned. Only affects actors with the Enemy tag. Added extra functionality to avoid collisions with the player character completely.

Added Checkpoint functionality, which is in the form of a checkpoint template. Which upon overlap collision, will transport the player character to a previously set vector location variable and set it's new spawn point.

To work with the checkpoint system, made a custom respawn method that resets the level but maintains the previously set checkpoint.